requirements.txt
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distro==1.9.0
dnspython==2.7.0
ecdsa>=0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
jwcrypto==1.5.6
openai==1.66.3
packaging==24.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.4.0
python-jwt==4.1.0
python-multipart==0.0.20
pytz==2025.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
qdrant-client==1.13.3
structlog==25.2.0
python-dateutil==2.9.0.post0
aio-pika==9.5.5
google-auth==2.40.1
google-generativeai==0.8.5
sendgrid===6.12.3
google-cloud-secret-manager===2.24.0
stripe
